More consistent error handling in GetSubnets
Change GetSubnets to return nil error at the end.

Before this change, GetSubnets would return the error return from
unmarshalling the last entry it got from etcd.  Unmarshalling errors are
logged anyway, and it is arbitrary that the last entry and none of the
other entries determine the error return from GetSubnets.